move humanReadableTimeZoneOffsetFormat to utils
